+++ api/v1alpha1/instancegroup_types.go
@@ -649,6 +649,21 @@ func (ig *InstanceGroup) Validate() error {
 		return errors.Errorf("validation failed, provisioner '%v' is invalid", s.Provisioner)
 	}
 
+	switch s.Provisioner {
+	case EKSProvisionerName:
+		if ig.GetEKSSpec() == nil {
+			return errors.Errorf("validation failed, provisioner '%v' not provided in spec", s.Provisioner)
+		}
+	case EKSManagedProvisionerName:
+		if ig.GetEKSManagedSpec() == nil {
+			return errors.Errorf("validation failed, provisioner '%v' not provided in spec", s.Provisioner)
+		}
+	case EKSFargateProvisionerName:
+		if ig.GetEKSFargateSpec() == nil {
+			return errors.Errorf("validation failed, provisioner '%v' not provided in spec", s.Provisioner)
+		}
+	}
+
 	if strings.EqualFold(s.Provisioner, EKSFargateProvisionerName) {
 		if err := s.EKSFargateSpec.Validate(); err != nil {
 			return err
@@ -1120,7 +1135,7 @@ func (ig *InstanceGroup) GetState() ReconcileState {
 
 func (ig *InstanceGroup) SetState(s ReconcileState) {
 	log.Info("state transition occured",
-		"instancegroup", ig.GetName(),
+		"instancegroup", ig.NamespacedName(),
 		"state", s,
 		"previousState", ig.Status.CurrentState,
 	)