+++ controllers/provisioners/eksmanaged/eksmanaged.go
@@ -34,6 +34,7 @@ const (
 )
 
 func (ctx *EksManagedInstanceGroupContext) CloudDiscovery() error {
+	ctx.processParameters()
 	var (
 		provisioned     = ctx.AwsWorker.IsNodeGroupExist()
 		discoveredState = ctx.GetDiscoveredState()
@@ -129,7 +130,7 @@ func (ctx *EksManagedInstanceGroupContext) Create() error {
 	if err != nil {
 		return err
 	}
-	ctx.Log.Info("created managed node group", "instancegroup", instanceGroup.GetName())
+	ctx.Log.Info("created managed node group", "instancegroup", instanceGroup.NamespacedName())
 	instanceGroup.SetState(v1alpha1.ReconcileModifying)
 	return nil
 }
@@ -176,7 +177,7 @@ func (ctx *EksManagedInstanceGroupContext) Update() error {
 		if err != nil {
 			return err
 		}
-		ctx.Log.Info("updated managed node group", "instancegroup", instanceGroup.GetName())
+		ctx.Log.Info("updated managed node group", "instancegroup", instanceGroup.NamespacedName())
 		instanceGroup.SetState(v1alpha1.ReconcileModifying)
 	} else {
 		instanceGroup.SetState(v1alpha1.ReconcileModified)
@@ -193,7 +194,7 @@ func (ctx *EksManagedInstanceGroupContext) Delete() error {
 	if err != nil {
 		return err
 	}
-	ctx.Log.Info("deleted managed node group", "instancegroup", instanceGroup.GetName())
+	ctx.Log.Info("deleted managed node group", "instancegroup", instanceGroup.NamespacedName())
 	instanceGroup.SetState(v1alpha1.ReconcileDeleting)
 	return nil
 }
@@ -229,8 +230,6 @@ func New(p provisioners.ProvisionerInput) *EksManagedInstanceGroupContext {
 	instanceGroup := ctx.GetInstanceGroup()
 
 	instanceGroup.SetState(v1alpha1.ReconcileInit)
-	ctx.processParameters()
-
 	return ctx
 }