+++ main.go
@@ -42,7 +42,7 @@ var (
 	setupLog = ctrl.Log.WithName("setup")
 )
 
-const controllerVersion = "instancemgr-0.10.0"
+const controllerVersion = "instancemgr-0.10.1"
 
 func init() {
 	instancemgrv1alpha1.AddToScheme(scheme)
@@ -114,7 +114,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	cacheCfg := cache.NewConfig(aws.CacheDefaultTTL, aws.CacheMaxItems, aws.CacheItemsToPrune)
+	cacheCfg := cache.NewConfig(aws.CacheDefaultTTL, aws.CacheBackgroundPruningInterval, aws.CacheMaxItems, aws.CacheItemsToPrune)
 
 	awsWorker := aws.AwsWorker{
 		Ec2Client: aws.GetAwsEc2Client(awsRegion, cacheCfg, maxAPIRetries),